refactor(types): изменить струтуру типов
