+++ src/client/analytics/requestTypes/futoi.ts
@@ -6,7 +6,7 @@ export interface IGetFutoiArgs {
 	/** Дата в формате YYYY-MM-DD, или "today" */
 	date?: "today" | string;
 	/** Последний срез за день, "1" - включить */
-	latest?: "0" | "1";
+	latest?: 0 | 1;
 	table_type?: "full" | string;
 }
