+++ src/client/info/requestTypes/events.ts
@@ -1,0 +1,8 @@
+import { AxiosResponse } from "axios";
+
+import { IEventsResponse } from "../responseTypes";
+
+export type TGetEvents = (
+	/** смещение начала */
+	start?: number,
+) => Promise<AxiosResponse<IEventsResponse>>;
