+++ src/client/info/requestTypes/eventsDetail.ts
@@ -1,0 +1,8 @@
+import { AxiosResponse } from "axios";
+
+import { IEventsDetailResponse } from "../responseTypes";
+
+export type TGetEventsDetail = (
+	/** номер события */
+	eventId: number,
+) => Promise<AxiosResponse<IEventsDetailResponse>>;
