+++ src/client/info/requestTypes/handbooks.ts
@@ -1,0 +1,5 @@
+import { AxiosResponse } from "axios";
+
+import { IHandbooksResponse } from "../responseTypes";
+
+export type TGetHandbooks = () => Promise<AxiosResponse<IHandbooksResponse>>;
