+++ src/client/info/requestTypes/index.ts
@@ -1,0 +1,7 @@
+export * from "./events";
+export * from "./eventsDetail";
+export * from "./handbooks";
+export * from "./news";
+export * from "./newsDetail";
+export * from "./turnovers";
+export * from "./turnoversColumns";
