+++ src/client/info/requestTypes/news.ts
@@ -1,0 +1,8 @@
+import { AxiosResponse } from "axios";
+
+import { INewsResponse } from "../responseTypes";
+
+export type TGetNews = (
+	/** смещение начала */
+	start?: number,
+) => Promise<AxiosResponse<INewsResponse>>;
