+++ src/client/info/requestTypes/newsDetail.ts
@@ -1,0 +1,8 @@
+import { AxiosResponse } from "axios";
+
+import { INewsDetailResponse } from "../responseTypes";
+
+export type TGetNewsDetail = (
+	/** номер новости */
+	newsId: number,
+) => Promise<AxiosResponse<INewsDetailResponse>>;
