+++ src/client/info/requestTypes/turnoversColumns.ts
@@ -1,0 +1,7 @@
+import { AxiosResponse } from "axios";
+
+import { ITurnoversColumnsResponse } from "../responseTypes";
+
+export type TGetTurnoversColumns = () => Promise<
+	AxiosResponse<ITurnoversColumnsResponse>
+>;
